requirements.txt
requests
geopy
python-dotenv
pyserial
crccheck